+++ NEWS.md
@@ -1,3 +1,4 @@
 # greta.distributions 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added `conditional_bernoulli` distribution (#5)